Move types into own file. Move functions from tests into utils.
