+++ debug.py
@@ -30,7 +30,6 @@ CONVOS = {
         "convo_id": "4kjz5rrrfe",
     },
 }
-from tests.utils.test_representativeness import expand_group_clusters_with_participants, generate_cluster_labels
 
 if True:
     # testing representativeness calculations
@@ -51,11 +50,11 @@ if True:
         cluster_labels = client.optimal_cluster_labels
     else:
         math_data = client.data_loader.math_data
-        group_clusters_with_pids = expand_group_clusters_with_participants(
+        group_clusters_with_pids = utils.expand_group_clusters_with_participants(
             group_clusters=math_data["group-clusters"],
             base_clusters=math_data["base-clusters"],
         )
-        cluster_labels = generate_cluster_labels(group_clusters_with_pids)
+        cluster_labels = utils.generate_cluster_labels(group_clusters_with_pids)
         client.optimal_cluster_labels = cluster_labels
 
     stats_by_group = utils.calculate_comment_statistics_by_group(
